# openapi.yaml
openapi: 3.0.3
info:
  title: Transfer Service API
  version: 1.1.0
  description: |
    API contract for a cross-platform file transfer tool.
    Auth model: JWT bearer token.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/shares:
    post:
      tags: [Shares]
      summary: Create a share link for a file or folder
      operationId: createShare
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareRequest'
      responses:
        '201':
          description: Share created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/s/{token}:
    get:
      tags: [Shares]
      summary: Get share metadata by token (public, no auth)
      operationId: getShareInfo
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Share metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareMeta'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Share link has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/s/{token}/download:
    get:
      tags: [Shares]
      summary: Download shared content by token (public, no auth)
      operationId: downloadShare
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File stream (single file) or zip archive (folder)
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Share link has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Folder download not yet implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateShareRequest:
      type: object
      required: [itemType, itemId]
      properties:
        itemType:
          type: string
          enum: [file, folder]
          example: file
        itemId:
          type: string
          description: ID of the file or folder to share
          example: f_01HZXA7J9QW2M3P4N5R6
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Optional expiry time; null means never expires
          example: "2026-04-01T00:00:00Z"
    ShareInfo:
      type: object
      required: [token, shareUrl, itemType, itemId, createdAt]
      properties:
        token:
          type: string
          example: aB3kR7mXpQ2wZv
        shareUrl:
          type: string
          description: Full public URL to access the share
          example: http://your-host/api/s/aB3kR7mXpQ2wZv
        itemType:
          type: string
          enum: [file, folder]
        itemId:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ShareMeta:
      type: object
      required: [token, itemType, name, createdAt]
      properties:
        token:
          type: string
        itemType:
          type: string
          enum: [file, folder]
        name:
          type: string
          description: Display name of the shared item
          example: project-plan.pdf
        size:
          type: integer
          format: int64
          nullable: true
          description: File size in bytes (null for folders)
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_CREDENTIALS
        message:
          type: string
          example: Username or password is incorrect.
        requestId:
          type: string
          example: req_01J2AA3M8T4D2Q9M
  responses:
    BadRequest:
      description: Invalid request payload/params
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: BAD_REQUEST
            message: Invalid request.
    Unauthorized:
      description: Missing or invalid token / wrong credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Authentication required.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FILE_NOT_FOUND
            message: File not found.
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Unexpected server error.